Update TestAll for Python 3 and Windows
Previously, the test file was designed to be run from the
directory where the file was located, but that only worked if the
module had already been installed to the system path. This was
fixed by setting up a `FILE_PATH` constant which always points at
the directory where the test file is located, no matter where the
test script is executed from.

And because `os.path.abspath` doesn't normalize forward slashes in
Windows, `os.path.join` is being used to normalize the path for
importing the other tests.