Fixed absolute path to XSD file
Previously this script would not properly strip out all of the
new references to the existing XSD file. Since these are not
actually needed, we are still stripping out most of the meta lines
which are never referenced again. For the `Location` definition,
which cannot be completely stripped out, we are just replacing the
location to the XSD file with a blank string.

This also removes the part where the Paypal field was normalized, as
this is no longer needed.